+++ zerver/tests/test_user_groups.py
@@ -26,6 +26,7 @@ from zerver.actions.user_groups import (
     do_change_user_group_permission_setting,
     do_deactivate_user_group,
     promote_new_full_members,
+    remove_subgroups_from_user_group,
 )
 from zerver.actions.users import do_deactivate_user
 from zerver.lib.create_user import create_user
@@ -529,6 +530,47 @@ class UserGroupAPITestCase(UserGroupTestCase):
         result = self.client_post("/json/user_groups/create", info=params)
         self.assert_json_error(result, "User group name cannot start with 'channel:'.")
         self.assert_length(NamedUserGroup.objects.filter(realm=hamlet.realm), 10)
+
+    def test_creating_groups_with_subgroups(self) -> None:
+        realm = get_realm("zulip")
+        hamlet = self.example_user("hamlet")
+        subgroup = check_add_user_group(realm, "support", [hamlet], acting_user=hamlet)
+        self.login("desdemona")
+
+        othello = self.example_user("othello")
+        params = {
+            "name": "Troubleshooting",
+            "members": orjson.dumps([othello.id]).decode(),
+            "description": "Troubleshooting team",
+            "subgroups": orjson.dumps([subgroup.id]).decode(),
+        }
+        result = self.client_post("/json/user_groups/create", info=params)
+        self.assert_json_success(result)
+        self.assert_length(NamedUserGroup.objects.filter(realm=hamlet.realm), 11)
+        user_group = NamedUserGroup.objects.get(name="Troubleshooting", realm=hamlet.realm)
+        self.assert_subgroup_membership(user_group, [subgroup])
+
+        # User can add subgroups to a group while creating it even if
+        # settings are set to not allow adding subgroups after creating
+        # the group.
+        self.login("othello")
+        self.assertEqual(realm.can_manage_all_groups.named_user_group.name, SystemGroups.OWNERS)
+
+        admins_group = NamedUserGroup.objects.get(
+            name=SystemGroups.ADMINISTRATORS, realm=realm, is_system_group=True
+        )
+        params = {
+            "name": "Backend",
+            "members": orjson.dumps([othello.id]).decode(),
+            "description": "Backend team",
+            "subgroups": orjson.dumps([subgroup.id]).decode(),
+            "can_manage_group": orjson.dumps(admins_group.id).decode(),
+            "can_add_members_group": orjson.dumps(admins_group.id).decode(),
+        }
+        result = self.client_post("/json/user_groups/create", info=params)
+        self.assert_json_success(result)
+        user_group = NamedUserGroup.objects.get(name="Troubleshooting", realm=realm)
+        self.assert_subgroup_membership(user_group, [subgroup])
 
     def do_test_set_group_setting_during_user_group_creation(self, setting_name: str) -> None:
         self.login("hamlet")
@@ -1671,9 +1713,27 @@ class UserGroupAPITestCase(UserGroupTestCase):
         # No notification message is sent for removing from user group.
         self.assertEqual(self.get_last_message(), initial_last_message)
 
+        # Test adding and removing subgroups.
+        admins_group = NamedUserGroup.objects.get(
+            name=SystemGroups.ADMINISTRATORS, realm=hamlet.realm, is_system_group=True
+        )
+        cordelia = self.example_user("cordelia")
+        subgroup = check_add_user_group(
+            hamlet.realm, "leadership", [cordelia], acting_user=cordelia
+        )
+        params = {"add_subgroups": orjson.dumps([subgroup.id, admins_group.id]).decode()}
+        result = self.client_post(f"/json/user_groups/{user_group.id}/members", info=params)
+        self.assert_json_success(result)
+        self.assert_subgroup_membership(user_group, [subgroup, admins_group])
+
+        params = {"delete_subgroups": orjson.dumps([admins_group.id]).decode()}
+        result = self.client_post(f"/json/user_groups/{user_group.id}/members", info=params)
+        self.assert_json_success(result)
+        self.assert_subgroup_membership(user_group, [subgroup])
+
         # Test when nothing is provided
         result = self.client_post(f"/json/user_groups/{user_group.id}/members", info={})
-        msg = 'Nothing to do. Specify at least one of "add" or "delete".'
+        msg = 'Nothing to do. Specify at least one of "add", "delete", "add_subgroups" or "delete_subgroups".'
         self.assert_json_error(result, msg)
         self.assert_user_membership(user_group, [hamlet])
 
@@ -2745,14 +2805,6 @@ class UserGroupAPITestCase(UserGroupTestCase):
         hamlet = self.example_user("hamlet")
         othello = self.example_user("othello")
 
-        moderators_group = NamedUserGroup.objects.get(name=SystemGroups.MODERATORS, realm=realm)
-        do_change_realm_permission_group_setting(
-            realm,
-            "can_manage_all_groups",
-            moderators_group,
-            acting_user=None,
-        )
-
         leadership_group = check_add_user_group(
             realm, "leadership", [desdemona, iago, hamlet], acting_user=desdemona
         )
@@ -2761,35 +2813,8 @@ class UserGroupAPITestCase(UserGroupTestCase):
         )
         test_group = check_add_user_group(realm, "test", [hamlet], acting_user=hamlet)
 
-        self.login("cordelia")
-        # Non-admin and non-moderators who are not a member of group cannot add or remove subgroups.
-        params = {"add": orjson.dumps([leadership_group.id]).decode()}
-        result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info=params)
-        self.assert_json_error(result, "Insufficient permission")
-
-        self.login("iago")
-        result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info=params)
-        self.assert_json_success(result)
-        self.assert_subgroup_membership(support_group, [leadership_group])
-
-        params = {"delete": orjson.dumps([leadership_group.id]).decode()}
-        result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info=params)
-        self.assert_json_success(result)
-        self.assert_subgroup_membership(support_group, [])
-
-        self.login("shiva")
-        params = {"add": orjson.dumps([leadership_group.id]).decode()}
-        result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info=params)
-        self.assert_json_success(result)
-        self.assert_subgroup_membership(support_group, [leadership_group])
-
-        params = {"delete": orjson.dumps([leadership_group.id]).decode()}
-        result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info=params)
-        self.assert_json_success(result)
-        self.assert_subgroup_membership(support_group, [])
-
         self.login("hamlet")
-        # Group creator can add or remove subgroups.
+        # Group creator can add or remove subgroups as they are member of can_manage_group.
         params = {"add": orjson.dumps([leadership_group.id]).decode()}
         result = self.client_post(f"/json/user_groups/{support_group.id}/subgroups", info=params)
         self.assert_json_success(result)
@@ -2818,7 +2843,7 @@ class UserGroupAPITestCase(UserGroupTestCase):
         )
         self.assert_subgroup_membership(support_group, [leadership_group])
 
-        self.login("iago")
+        self.login("desdemona")
         params = {"add": orjson.dumps([support_group.id]).decode()}
         result = self.client_post(f"/json/user_groups/{leadership_group.id}/subgroups", info=params)
         self.assert_json_error(
@@ -2884,6 +2909,263 @@ class UserGroupAPITestCase(UserGroupTestCase):
         params = {"add": orjson.dumps([support_group.id]).decode()}
         result = self.client_post(f"/json/user_groups/{test_group.id}/subgroups", info=params)
         self.assert_json_error(result, "User group is deactivated.")
+
+    def test_permission_to_add_subgroups_to_group(self) -> None:
+        realm = get_realm("zulip")
+        hamlet = self.example_user("hamlet")
+        othello = self.example_user("othello")
+        leadership_group = check_add_user_group(realm, "leadership", [othello], acting_user=othello)
+        support_group = check_add_user_group(realm, "support", [hamlet], acting_user=hamlet)
+
+        def check_adding_subgroups_to_group(acting_user: str, error_msg: str | None = None) -> None:
+            params = {"add": orjson.dumps([leadership_group.id]).decode()}
+            self.assert_subgroup_membership(support_group, [])
+
+            result = self.api_post(
+                self.example_user(acting_user),
+                f"/api/v1/user_groups/{support_group.id}/subgroups",
+                info=params,
+            )
+            if error_msg is None:
+                self.assert_json_success(result)
+                self.assert_subgroup_membership(support_group, [leadership_group])
+                # Remove the subgroup to test further cases.
+                remove_subgroups_from_user_group(
+                    support_group, [leadership_group], acting_user=None
+                )
+            else:
+                self.assert_json_error(result, error_msg)
+
+        nobody_group = NamedUserGroup.objects.get(
+            name=SystemGroups.NOBODY, realm=realm, is_system_group=True
+        )
+        # Set manage permissions to "Nobody" group to test permission
+        # with can_add_members_group.
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_manage_group",
+            nobody_group,
+            acting_user=None,
+        )
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            nobody_group,
+            acting_user=None,
+        )
+
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_add_members_group",
+            nobody_group,
+            acting_user=None,
+        )
+        check_adding_subgroups_to_group("desdemona", "Insufficient permission")
+
+        owners_group = NamedUserGroup.objects.get(
+            name=SystemGroups.OWNERS, realm=realm, is_system_group=True
+        )
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_add_members_group",
+            owners_group,
+            acting_user=None,
+        )
+        check_adding_subgroups_to_group("iago", "Insufficient permission")
+        check_adding_subgroups_to_group("desdemona")
+
+        # Test case when setting is set to a non-system group.
+        prospero = self.example_user("prospero")
+        test_group = check_add_user_group(realm, "test", [prospero], acting_user=prospero)
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_add_members_group",
+            test_group,
+            acting_user=None,
+        )
+        check_adding_subgroups_to_group("desdemona", "Insufficient permission")
+        check_adding_subgroups_to_group("prospero")
+
+        # Test case when setting is set to an anonymous group.
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            direct_members=[othello],
+            direct_subgroups=[owners_group],
+        )
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_add_members_group",
+            setting_group,
+            acting_user=None,
+        )
+        check_adding_subgroups_to_group("prospero", "Insufficient permission")
+        check_adding_subgroups_to_group("iago", "Insufficient permission")
+        check_adding_subgroups_to_group("desdemona")
+        check_adding_subgroups_to_group("othello")
+
+        # Set can_add_members_group setting to nobody, so we can test
+        # managing permissions as well.
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_add_members_group",
+            nobody_group,
+            acting_user=None,
+        )
+
+        # Check permission as per can_manage_group setting.
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            direct_members=[othello],
+            direct_subgroups=[owners_group],
+        )
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_manage_group",
+            setting_group,
+            acting_user=None,
+        )
+        check_adding_subgroups_to_group("iago", "Insufficient permission")
+        check_adding_subgroups_to_group("desdemona")
+        check_adding_subgroups_to_group("othello")
+
+        # Check permission as per can_manage_all_groups setting.
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_manage_group",
+            nobody_group,
+            acting_user=None,
+        )
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            direct_members=[othello],
+            direct_subgroups=[owners_group],
+        )
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            setting_group,
+            acting_user=None,
+        )
+        check_adding_subgroups_to_group("iago", "Insufficient permission")
+        check_adding_subgroups_to_group("desdemona")
+        check_adding_subgroups_to_group("othello")
+
+    def test_permission_to_remove_subgroups_from_group(self) -> None:
+        realm = get_realm("zulip")
+        hamlet = self.example_user("hamlet")
+        othello = self.example_user("othello")
+        leadership_group = check_add_user_group(realm, "leadership", [othello], acting_user=othello)
+        support_group = check_add_user_group(realm, "support", [hamlet], acting_user=hamlet)
+        add_subgroups_to_user_group(support_group, [leadership_group], acting_user=None)
+
+        def check_remove_subgroups_from_group(
+            acting_user: str, error_msg: str | None = None
+        ) -> None:
+            params = {"delete": orjson.dumps([leadership_group.id]).decode()}
+            self.assert_subgroup_membership(support_group, [leadership_group])
+
+            result = self.api_post(
+                self.example_user(acting_user),
+                f"/api/v1/user_groups/{support_group.id}/subgroups",
+                info=params,
+            )
+            if error_msg is None:
+                self.assert_json_success(result)
+                self.assert_subgroup_membership(support_group, [])
+                # Add the subgroup again to test further cases.
+                add_subgroups_to_user_group(support_group, [leadership_group], acting_user=None)
+            else:
+                self.assert_json_error(result, error_msg)
+
+        # Set permissions for managing all groups to "Nobody" group to
+        # test permission with can_manage_group.
+        nobody_group = NamedUserGroup.objects.get(
+            name=SystemGroups.NOBODY, realm=realm, is_system_group=True
+        )
+        do_change_realm_permission_group_setting(
+            realm, "can_manage_all_groups", nobody_group, acting_user=None
+        )
+
+        owners_group = NamedUserGroup.objects.get(
+            name=SystemGroups.OWNERS, realm=realm, is_system_group=True
+        )
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_manage_group",
+            owners_group,
+            acting_user=None,
+        )
+        check_remove_subgroups_from_group("iago", "Insufficient permission")
+        check_remove_subgroups_from_group("desdemona")
+
+        # Test case when setting is set to a non-system group.
+        prospero = self.example_user("prospero")
+        test_group = check_add_user_group(realm, "test", [prospero], acting_user=prospero)
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_manage_group",
+            test_group,
+            acting_user=None,
+        )
+        check_remove_subgroups_from_group("desdemona", "Insufficient permission")
+        check_remove_subgroups_from_group("prospero")
+
+        # Test case when setting is set to an anonymous group.
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            direct_members=[othello],
+            direct_subgroups=[owners_group],
+        )
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_manage_group",
+            setting_group,
+            acting_user=None,
+        )
+        check_remove_subgroups_from_group("prospero", "Insufficient permission")
+        check_remove_subgroups_from_group("iago", "Insufficient permission")
+        check_remove_subgroups_from_group("desdemona")
+        check_remove_subgroups_from_group("othello")
+
+        # Set can_manage_group setting to nobody, so we can test
+        # can_manage_all_groups behavior.
+        do_change_user_group_permission_setting(
+            support_group,
+            "can_manage_group",
+            nobody_group,
+            acting_user=None,
+        )
+
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            owners_group,
+            acting_user=None,
+        )
+        check_remove_subgroups_from_group("desdemona")
+        check_remove_subgroups_from_group("iago", "Insufficient permission")
+
+        # Test case when setting is set to a non-system group.
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            test_group,
+            acting_user=None,
+        )
+        check_remove_subgroups_from_group("desdemona", "Insufficient permission")
+        check_remove_subgroups_from_group("prospero")
+
+        # Test case when setting is set to an anonymous group.
+        setting_group = self.create_or_update_anonymous_group_for_setting(
+            direct_members=[othello],
+            direct_subgroups=[owners_group],
+        )
+        do_change_realm_permission_group_setting(
+            realm,
+            "can_manage_all_groups",
+            setting_group,
+            acting_user=None,
+        )
+        check_remove_subgroups_from_group("prospero", "Insufficient permission")
+        check_remove_subgroups_from_group("iago", "Insufficient permission")
+        check_remove_subgroups_from_group("desdemona")
+        check_remove_subgroups_from_group("othello")
 
     def test_get_is_user_group_member_status(self) -> None:
         self.login("iago")